use sim cfg in sim dataset
